Updated .gitignore to better reflect the kind of files generated by Android Studio that we don't want to check in.
Updated build.gradle so it can do a build that is just stored in the local Maven (just call the install task), great for development! Also made UPLOAD_VERSION_CBLITE into a property instead of an environment variable.

Enhanced CBLRouter so it will accept a CBLRequestAuthorization and check it.

Enhanced CBLURLConnection so one can set/get a SSLSession for it, need for SSL authorization validation.

Checking in the output files from gradle wrapper so we all build in the same way

Added CBLRequestAuthorization interface to handle authorization calls.