check if pulled revisions are deleted by inspecting their properties
The code was relying on the "deleted" property in a _changes feed entry
to be "true" for all deleted revisions, but this isn't true in CouchDB,
which might return live and deleted revisions together in one entry.

For example, in the change entry

  {"seq":2,"id":"d5abe6fe-c3c8-4592-b0c1-eefa20870794","changes":[{"rev":"3-9a5884fe9144ed948db9ff983dbfcae9"},{"rev":"3-8bdfaebacbc21a916633103c74f17c52"}]}

the first revision may be live, while the second is deleted. So, the
actual body of the revision has to be checked too.