Improving logging on retry failures
I am running requests over TOR hidden services that can fail for very long times. So I need to track when we have a failure that causes a remote request to be given up on. Is there anywhere else a logging even or notification would occur if the replication couldn't get a connection established?