Make setting the logger level actually work
Right now when we set the log level via the Couchbase logger it will (in Java anyway) bottom out into the the SystemLogger class. But the logger object in this class has its own logging level independent of the one we set in Couchbase and that level by default is 'info'. So it doesn't matter if you set the Couchbase logging level lower, logger won't let it through. The fix is to just set the logger's level to ALL and just depend on Couchbase's logger to handle filtering.