This clean up isn't necessary anymore
This code was originally taken from stop in order to prevent multiple
threads from needing to interact with the request object. But the way the
code works this isn't necessary since when stop is called it will be
honored at the top of the while loop and by then request will have been
handled one way or another.