BlobStoreWriter: Throw IOExceptions.
Instead of throwing an unchecked exception and ignoring the close on the
output file, throw checked IOExceptions.

We should consider an fsync before closing the outputstream to get a better
consistency guarantee.

With this commit, when saving an attachment fails on UnsavedRevison#save
a CBLException is thrown with STATUS_ATTACHMENT_ERROR.

MultipartDocumentReader continues to throw unchecked exceptions.