Serve attachments unzipped.
When retrieving an attachment through a Revision object, the client
does not expect a gzipped attachment because getGzipped() is not part
of the "public" API and is also left unset if the stream is gzipped.

Text attachments are normally replicated gzip encoded, so users are left
to either probe for gzip encoding or just blindly wrapping the stream
in a GZIPInputStream and hope they notice when this behaviour is fixed.

This commit only affects Attachment's retrieved through revisions (eg.
those that lazilly retrieve the stream) and does not affect Router, which
continue to get them with the original encoding.

It should be noted that .NET CBL is already doing this (not sure about iOS).