Fixed dead lock when calling Batcher.flushAll()
PullerInternal calls Batcher.flushAll() from the same work executor given to the Batcher. Now make Batcher.flushAll() with an option to wait or not to wait for all objects to be processed.

Also make sure that while executing flushAll(), all new queued objects will be scheduled to work by flushAll().