Fix the single shot pull replication to pull up to 50 docs
The limit was set to 50 in the code for some time but has never been used. In 1.3, the code that apply the limit (if the limit is set) was added.

Adding the limit caused the issue that a single shot pull replication will only pull up to the limit and stop. There is no limit set in couchbase-lite-ios as well. Simply removing the limit should solve the issue.

https://github.com/couchbase/couchbase-lite-android/issues/955