Implement GET _changes feed timeout to prevent dead connection leaks
Implemented GET _changes feed timeout to prevent dead connection leaks. The implementation is inline with couchdb and SG. The default timeout is 60000 miliseconds (same as couchdb but different from SG which has a default timeout as 5 mins) and the timeout will be disabled if heartbeat is specified.

Note that couchbase-lite-ios doesn’t implement the timeout feature as the iOS listener doesn’t seem to have half-close socket issues as couchbase-lite-android is having.

#1296