Fix cast attachment length
From #1379, jackson sometimes deserializes number into float instead of int. Instead of casting the value into Integer directly, cast it to Number and then get the integer / long value from the Number object. This changes make the code consistent with the other places in couchbase-java-core when getting the attachment length.