Enable TLSv1.1 and v1.2 on Android API 16 - 19
Based on https://developer.android.com/reference/javax/net/ssl/SSLSocket, Android API 16 - 19 have already supported TLSv1.1 and 1.2 but do not enable them by default.

This commit adds a workaround to create a TLSSocketFactory which is a wrapper of a TLS based SSLSocketFactory to enable TLSv1.1 and TSL1.2. This is enabled only on Android but not on Java.

https://github.com/couchbase/couchbase-lite-android/issues/1635